Set TerminationMessagePolicy on Initializer to use logs as fallback
While not explicitly used, with the removal of the log fetching from the
Initializer (Job->Pod->Container) any potential source for verbose information
about the error might be lost. The Kubernetes approach to provide more details
than the exit code about the reason a container terminated is the termination
message (see [1]).

Per default this message needs to be explictly written to `/dev/termination-log`,
but there also is the option to use the last bit of log output as fallback source.

[1] https://kubernetes.io/docs/tasks/debug/debug-application/determine-reason-pod-failure/